Move fixture to conftest file
Moving fixture of graph from test to conftest file to be able to share it
with multiple files.
